fix(wip): use MPS when available except for sparse computations
Selecting MPS as a backend when CUDA is unavailable does not initially work
properly, as the code uses sparse tensors and the SparseMPS backend is not
implemented. This commit changes the code so that when MPS is in use, sparse
tensors are created on the CPU instead, and then moved back to the same device
as the rest of the computation. Since Apple Silicon has unified memory, this
"movement" should be an accounting change only and a no-op in terms of actual
data movement.

I haven't fully tested this yet, but it provides a significant performance
improvement on an M3 MBA, taking the time to execute all cells in the
circuit_tracing_tutorial.ipynb from ~19min to <1min.